+++ scripts/installation/PutBug.sh
@@ -1,0 +1,3 @@
+#! /bin/bash
+cd /home/ws/php/laraknife/scripts/installation
+rsync -av CreateLaraProj.sh IncludeLara.sh hm@bug.hamatoma.de:/home/www/public/public