+++ scripts/laraknife-tool.sh
@@ -2,12 +2,29 @@
 
 MODE=$1
 PROJ=$(basename $(pwd))
+# ===
 function Usage(){
-  echo "Usage: laraknife-tool.sh build-links [[--force]]}"
-  echo "  run it from the root directory of the composer package"
-  echo "  --force: existing links will be removed before"
+  cat <<EOS
+Usage: laraknife-tool.sh TASK
+  Run it from the root directory of the composer package
+<TASKS>:
+  build-links [--force]
+    Creates links for unchangable files in laraknife
+    --force: existing links will be removed before
+  init-i18n
+    Setup I18N
+  fill-db
+    Puts records into tables
+  adapt-modules
+    Integrates the laraknife modules into the project
+  create-layout
+    Creates a layout blade file
+  create-home
+    Creates a homepage
+EOS
   echo "+++ $*"
 }
+# ===
 function BuildLinks(){
   local option="$1"
   local base=vendor/hamatoma/laraknife
@@ -41,7 +58,7 @@ function BuildLinks(){
     for resource in css js; do
       mkdir -p public/$resource
       test "$option" = "--force" && rm -fv public/$resource/laraknife.$resource
-      ln -s ../../$dirResources/$resource/laraknife.$resource public/$resource/laraknife.$resource
+      ln -s ../../$dirResources/$resource/laraknife.$resource public/$resource/laraknife.$resource public/$resource
     done
     local fn=public/css/$PROJ.css
     cat <<EOS >$fn
@@ -64,8 +81,6 @@ input, button {
 }
 EOS
     echo "= created: $fn"
-    test "$option" = "--force" && rm -fv resources/views/$module
-    ln -s ../../../$dirResources/js/laraknife.js
     # === DB creation
     local node=''
     for file in $dirTemplates/database/migrations/*.php; do
@@ -103,12 +118,112 @@ EOS
 php app/Helpers/Builder.php $*
 EOS
   chmod +x $script
+  test "$option" = "--force" && rm larascripts
+  ln -s $base/scripts larascripts
+}
+# ===
+function InitI18N(){
+  local fn=config/app.php
+  if [ ! -f $fn ]; then
+    echo "++ missing $fn"
+  else
+    sed -i -e "s/'locale' => 'en'/'locale' => 'de_DE',#N#'available_locales' => [#N#  'English' => 'en',#N#  'German' => 'de_DE'#N#]/" \
+    -e 's/#N#/\n/g' $fn
+    grep -A4 "locale.*de_DE" $fn
+    local fn=resources/lang/sources/$PROJ.de.json
+    if [ ! -f $fn ]; then
+      cat <<EOS >$fn
+{
+"!comment": "Bitte alphabetisch sortiert eintragen",
+"ZZZZZ_last": ""
+}
+EOS
+      echo "created: $fn"
+      ls -l $fn
+    fi
+  fi
+}
+# ===
+function FillDb(){
+  php artisan migrate
+  sudo mysql lrv$PROJ <<'EOS'
+insert into roles (name, priority, created_at, updated_at) values 
+('Administrator', 10, '2023.12.28', '2023-12-28'),
+('Manager', 20, '2023.12.28', '2023-12-28'),
+('User', 30, '2023.12.28', '2023-12-28'),
+('Guest', 90, '2023.12.28', '2023-12-28');
+insert into sproperties (id, scope, name, `order`, shortname, created_at) values
+(1001, 'status', 'active', 10, 'A', '2023-12-28'),
+(1002, 'status', 'inactive', 20, 'I', '2023-12-28'),
+(2001, 'category', 'standard', 10, '-', '2023-12-28'),
+(2002, 'category', 'private', 20, 'P', '2023-12-28'),
+(2003, 'category', 'work', 30, 'W', '2023-12-28'),
+(2601, 'notestatus', 'open', 10, 'O', '2023-12-28'),
+(2602, 'notestatus', 'closed', 20, 'C', '2023-12-28');
+select count(*) from roles as role_count;
+select count(*) from sproperties as sproperty_count;
+EOS
+}
+# ===
+function AdaptModules(){
+  local fn=app/Models/User.php
+  found=$(grep role_id $fn)
+  if [ -n "$found" ]; then
+    echo "= role_id already found"
+  else
+    sed -i -e "s/'password'/'password',#N#    'role_id'/" -e 's/#N#/\n/g' $fn
+  fi
+  grep -A5 fillable $fn
+  fn=routes/web.php
+  found=$(grep RoleController::routes $fn)
+  if [ -n "$found" ]; then
+    echo "= RoleController::routes already found"
+  else
+    sed -i -e 's/Route;/Route;#N#use App\\Http\\Controllers\\RoleController;#N#use App\\Http\\Controllers\\UserController;#N#use App\\Http\\Controllers\\SPropertyController;/' \
+      -e 's/Auth::routes..;/Auth::routes();#N#RoleController::routes();#N#SPropertyController::routes();#N#UserController::routes();/' \
+      -e 's/#N#/\n/g' \
+      $fn
+    echo "= routes adapted:"
+  fi
+  grep -A3 RoleController::routes $fn
+}
+# ===
+function CreateLayout(){
+  . project.env
+  local fn=resources/views/layouts/$PROJ.blade.php
+  sed -e "s/PROJECT/$PROJ/g" vendor/hamatoma/laraknife/templates/layout.templ >$fn
+  cd resources/views/layouts
+  ln -sv $PROJ.blade.php backend.blade.php
+  cd ../../..
+  echo "= layout $fn and backend.blade.php have been created"
+}
+# ===
+function CreateHome(){
+ . project.env
+  local fn=resources/views/home.blade.php
+  sed -e "s/PROJECT/$PROJ/g" vendor/hamatoma/laraknife/templates/home.templ >$fn
+  echo "= home $fn has been created"
 }
 case $MODE in
 build-links)
   BuildLinks $2
   ;;
+init-i18n)
+  InitI18N
+  ;;
+fill-db)
+  FillDb
+  ;;
+adapt-modules)
+  AdaptModules
+  ;;
+create-layout)
+  CreateLayout
+  ;;
+create-home)
+  CreateHome
+  ;;
 *)
-  Usage "unknown MODE: $MODE"
+  Usage "unknown TASK: $MODE"
   ;;
 esac